Refine SlugGenerator package and update README
This commit removes unnecessary configurations in SlugGeneratorServiceProvider and updates SlugGenerator component to throw InvalidArgumentException when the defined slug field does not exist in the model attributes. It modifies the visibility of getSluggableField method from public to protected. The README has also been enhanced specifying how to use and configure the package.